# Faster IFeatureCollection.Get<TFeature>