… benchmarks/Kestrel.Performance/HttpProtocolFeatureCollection.cs
… test/Kestrel.Core.Tests/HttpProtocolFeatureCollectionTests.cs
… tools/CodeGenerator/HttpProtocolFeatureCollection.cs