Go via Generic rather than wrapping with `(TFeature)Get(typeof(TFeature))` as it scales better

pre
```
                 Method |     Mean |     Error |    StdDev |         Op/s | Allocated |
----------------------- |---------:|----------:|----------:|-------------:|----------:|
     GetViaTypeOf_First | 12.56 ns | 0.1568 ns | 0.2347 ns | 79,608,881.2 |       0 B |
    GetViaGeneric_First | 22.93 ns | 0.1275 ns | 0.1908 ns | 43,613,651.3 |       0 B |
      GetViaTypeOf_Last | 75.97 ns | 0.1121 ns | 0.1678 ns | 13,163,676.0 |       0 B |
     GetViaGeneric_Last | 85.30 ns | 0.2586 ns | 0.3870 ns | 11,723,594.4 |       0 B |
    GetViaTypeOf_Custom | 77.44 ns | 0.2239 ns | 0.3351 ns | 12,912,708.4 |       0 B |
   GetViaGeneric_Custom | 86.67 ns | 0.3066 ns | 0.4590 ns | 11,538,511.2 |       0 B |
  GetViaTypeOf_NotFound | 77.56 ns | 0.2899 ns | 0.4339 ns | 12,893,115.2 |       0 B |
 GetViaGeneric_NotFound | 86.23 ns | 0.2085 ns | 0.3121 ns | 11,597,065.2 |       0 B |
```
post
```
                 Method |     Mean |     Error |    StdDev |         Op/s | Allocated |
----------------------- |---------:|----------:|----------:|-------------:|----------:|
     GetViaTypeOf_First | 11.85 ns | 0.0565 ns | 0.0846 ns | 84,418,379.7 |       0 B |
    GetViaGeneric_First | 21.11 ns | 0.4077 ns | 0.6102 ns | 47,376,257.7 |       0 B |
      GetViaTypeOf_Last | 74.42 ns | 0.3818 ns | 0.5714 ns | 13,437,263.4 |       0 B |
     GetViaGeneric_Last | 40.41 ns | 4.6554 ns | 6.9680 ns | 24,748,132.8 |       0 B |
    GetViaTypeOf_Custom | 76.88 ns | 0.3606 ns | 0.5397 ns | 13,006,456.8 |       0 B |
   GetViaGeneric_Custom | 41.89 ns | 0.1967 ns | 0.2943 ns | 23,872,788.5 |       0 B |
  GetViaTypeOf_NotFound | 77.13 ns | 0.1864 ns | 0.2790 ns | 12,965,485.2 |       0 B |
 GetViaGeneric_NotFound | 41.31 ns | 0.1276 ns | 0.1910 ns | 24,204,614.8 |       0 B |
```

#2012 rebased

Resolves https://github.com/aspnet/KestrelHttpServer/issues/2303