@benaadams 2018-02-03
Similar output other than benchmarks only outputting a single column...
```
         Op/s |
-------------:|
 83,258,549.1 |
 48,740,215.0 |
 13,415,794.9 |
 25,695,864.4 |
 13,010,660.7 |
 24,729,265.6 |
 12,951,687.0 |
 25,009,785.9 |
```