@benaadams 2018-02-10
PR also shaves a significant amount off `OnReset()` 

52 secs of plaintext
Pre
![image](https://user-images.githubusercontent.com/1142958/36067100-5c023c82-0eae-11e8-8a94-1e0f734fc918.png)



Post
![image](https://user-images.githubusercontent.com/1142958/36067096-417fa2d2-0eae-11e8-856e-a53322bc6dd2.png)