@benaadams 2018-02-11
Set and Get branches

Pre
![image](https://user-images.githubusercontent.com/1142958/36067873-91b4f2a0-0ebf-11e8-8201-ee240803f040.png)

Post
![image](https://user-images.githubusercontent.com/1142958/36067876-b7858ba2-0ebf-11e8-82e9-4cc26397e577.png)