@benaadams 2018-02-17
> Any idea why GetViaGeneric_First is still twice as fast as GetViaGeneric_Last? Shouldn't the JIT remove all the extra branches so it optimizes to return `(ReifiedType)_specificFeature`?

The interfaces are object types so it produces a shared generic and doesn't branch eliminate; have raised issue https://github.com/dotnet/coreclr/issues/16266 requesting ability to mark a method and have it be generated as non-shared (though may be a large Jit change).

If the method inlined (size aside), it would eliminate all the branches; but alas its an interface method and you can't inline through an interface. 

Additionally an explicit interface method counts as virtual method https://github.com/dotnet/coreclr/issues/16198 so will only devirtualize on a sealed type where the explicit type is known/exposed at Jit time; so this can never devirtualize and inline (currently anyway):

```csharp
((IFeatureCollection)this).Set<IHttpUpgradeFeature>(this);
```

> Theoretically, isn't this exactly what `Set<IHttpUpgradeFeature>(this)` should optimize to?

To work around it you could define `protected FastFeatureSet<TFeature>(TFeature instance)` on `HttpProtocol` with an aggressive inline and hope it inlines into both two single feature resets (on `Http1Connection` and `Http2Connection`) and into the interface generic `void IFeatureCollection.Set<TFeature>(TFeature instance)`, or not define it explicitly `public Set<TFeature>(TFeature instance)`; else it introduces an extra method call and moves the *potential* interface inline one step further away; and crossgen isn't as happy with inlining shared generics with this number of type references (give error like "complex type handling"); the runtime Jit is happier to do it though.

It also seems a lot of work and hoops to jump through; as well as Jit beahviour to rely on just to avoid defining two simple Reset methods?